# query: introduce space optimized BranchesRepos