--- a/eval.go
+++ b/eval.go
@@ -69,14 +69,9 @@ func (d *indexData) simplify(in query.Q) query.Q {
 				return strings.Contains(repo.Name, r.Pattern)
 			})
 		case *query.BranchesRepos:
-			if len(d.repoMetaData) == 1 {
-				// Can simplify query now. compound too complicated since each repo
-				// may have different branches.
-				return r.Branches(d.repoMetaData[0].ID)
-			}
-			for _, md := range d.repoMetaData {
+			for i := range d.repoMetaData {
 				for _, br := range r.List {
-					if br.Repos.Contains(md.ID) {
+					if br.Repos.Contains(d.repoMetaData[i].ID) {
 						return q
 					}
 				}