--- a/matchtree.go
+++ b/matchtree.go
@@ -931,7 +931,7 @@ func (d *indexData) newMatchTree(q query.Q) (matchTree, error) {
 	case *query.RepoSet:
 		reposWant := make([]bool, len(d.repoMetaData))
 		for repoIdx, r := range d.repoMetaData {
-			if _, ok := s.Set[r.Name]; ok {
+			if s.Contains(r.Name, r.ID) {
 				reposWant[repoIdx] = true
 			}
 		}