--- a/query/marshal.go
+++ b/query/marshal.go
@@ -5,7 +5,10 @@ import (
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"io"
 	"unsafe"
+
+	"github.com/RoaringBitmap/roaring"
 )
 
 // We implement a custom binary marshaller for a list of repos to
@@ -48,61 +51,99 @@ import (
 // RepoBranches_Encode-8     20.0k ± 0%      0.0k ± 0%  -100.00%  (p=0.000 n=10+10)
 // RepoBranches_Decode-8     50.6k ± 0%      0.4k ± 0%   -99.26%  (p=0.000 n=10+10)
 
-// repoBranchesEncode implements an efficient encoder for RepoBranches.
-func repoBranchesEncode(repoBranches map[string][]string) ([]byte, error) {
+func (rb *RepoBranches) MarshalBinary() ([]byte, error) {
 	var b bytes.Buffer
 	var enc [binary.MaxVarintLen64]byte
-	varint := func(n int) {
-		m := binary.PutUvarint(enc[:], uint64(n))
+	varint := func(n uint64) {
+		m := binary.PutUvarint(enc[:], n)
 		b.Write(enc[:m])
 	}
 	str := func(s string) {
-		varint(len(s))
+		varint(uint64(len(s)))
 		b.WriteString(s)
 	}
-	strSize := func(s string) int {
-		return binary.PutUvarint(enc[:], uint64(len(s))) + len(s)
+	strSize := func(s string) uint64 {
+		return uint64(binary.PutUvarint(enc[:], uint64(len(s))) + len(s))
 	}
 
-	// Calculate size
-	size := 1 // version
-	size += binary.PutUvarint(enc[:], uint64(len(repoBranches)))
-	for name, branches := range repoBranches {
-		size += strSize(name) + 1
-		if l := len(branches); l == 1 && branches[0] == "HEAD" {
-			continue
-		} else if l == 0 {
-			// We reserve "0" for the "HEAD" special case.
-			return nil, fmt.Errorf("repo with no branches: %q", name)
-		} else if l > 255 {
-			// We encode branches len as a byte (saves 11% cpu vs varint). This is
-			// fine sinze Zoekt can only index upto 64 branches (uses a bitmask on a
-			// 64bit int to encode branch information for a document)
-			return nil, fmt.Errorf("can't encode more than 255 branches: %d", l)
-		}
-		for _, branch := range branches {
+	switch {
+	case len(rb.IDs) > 0:
+		// Calculate size
+		size := uint64(1) // version
+		size += uint64(binary.PutUvarint(enc[:], uint64(len(rb.IDs))))
+		for branch, ids := range rb.IDs {
 			size += strSize(branch)
-		}
-	}
-	b.Grow(size)
-
-	// Version
-	b.WriteByte(1)
-
-	// Length
-	varint(len(repoBranches))
-
-	for name, branches := range repoBranches {
-		str(name)
-
-		// Special case "HEAD"
-		if len(branches) == 1 && branches[0] == "HEAD" {
-			branches = nil
+			idsSize := ids.GetSerializedSizeInBytes()
+			size += uint64(binary.PutUvarint(enc[:], idsSize))
+			size += idsSize
 		}
 
-		b.WriteByte(byte(len(branches)))
-		for _, branch := range branches {
+		b.Grow(int(size))
+
+		// Version
+		b.WriteByte(2)
+
+		// Length
+		varint(uint64(len(rb.IDs)))
+
+		for branch, ids := range rb.IDs {
 			str(branch)
+			ids.RunOptimize()
+			l := ids.GetSerializedSizeInBytes()
+			varint(l)
+
+			n, err := ids.WriteTo(&b)
+			if err != nil {
+				return nil, err
+			}
+
+			if uint64(n) != l {
+				return nil, io.ErrShortWrite
+			}
+		}
+
+	case len(rb.Set) > 0:
+		// Calculate size
+		size := uint64(1) // version
+		size += uint64(binary.PutUvarint(enc[:], uint64(len(rb.Set))))
+		for name, branches := range rb.Set {
+			size += strSize(name) + 1
+			if l := len(branches); l == 1 && branches[0] == "HEAD" {
+				continue
+			} else if l == 0 {
+				// We reserve "0" for the "HEAD" special case.
+				return nil, fmt.Errorf("repo with no branches: %q", name)
+			} else if l > 255 {
+				// We encode branches len as a byte (saves 11% cpu vs varint). This is
+				// fine sinze Zoekt can only index upto 64 branches (uses a bitmask on a
+				// 64bit int to encode branch information for a document)
+				return nil, fmt.Errorf("can't encode more than 255 branches: %d", l)
+			}
+			for _, branch := range branches {
+				size += strSize(branch)
+			}
+		}
+
+		b.Grow(int(size))
+
+		// Version
+		b.WriteByte(1)
+
+		// Length
+		varint(uint64(len(rb.Set)))
+
+		for name, branches := range rb.Set {
+			str(name)
+
+			// Special case "HEAD"
+			if len(branches) == 1 && branches[0] == "HEAD" {
+				branches = nil
+			}
+
+			b.WriteByte(byte(len(branches)))
+			for _, branch := range branches {
+				str(branch)
+			}
 		}
 	}
 
@@ -114,78 +155,106 @@ func repoBranchesEncode(repoBranches map[string][]string) ([]byte, error) {
 // repoBranches slice, so it is safe to share this slice.
 var head = []string{"HEAD"}
 
-// repoBranchesDecode implements an efficient decoder for RepoBranches.
-func repoBranchesDecode(b []byte) (map[string][]string, error) {
+func (rb *RepoBranches) UnmarshalBinary(b []byte) error {
 	// binaryReader returns strings pointing into b to avoid allocations. We
 	// don't own b, so we create a copy of it.
 	r := binaryReader{b: append([]byte{}, b...)}
 
-	// Version
-	if v := r.byt(); v != 1 {
-		return nil, fmt.Errorf("unsupported RepoBranches encoding version %d", v)
-	}
+	switch v := r.byt(); v { // Version
+	case 1:
+		// Length
+		l := r.uvarint()
+		repoBranches := make(map[string][]string, l)
 
-	// Length
-	l := r.uvarint()
-	repoBranches := make(map[string][]string, l)
+		for i := 0; i < l; i++ {
+			name := r.str()
 
-	for i := 0; i < l; i++ {
-		name := r.str()
+			branchesLen := int(r.byt())
 
-		branchesLen := int(r.byt())
+			// Special case "HEAD"
+			if branchesLen == 0 {
+				repoBranches[name] = head
+				continue
+			}
 
-		// Special case "HEAD"
-		if branchesLen == 0 {
-			repoBranches[name] = head
-			continue
+			branches := make([]string, branchesLen)
+			for j := 0; j < branchesLen; j++ {
+				branches[j] = r.str()
+			}
+			repoBranches[name] = branches
 		}
 
-		branches := make([]string, branchesLen)
-		for j := 0; j < branchesLen; j++ {
-			branches[j] = r.str()
-		}
-		repoBranches[name] = branches
-	}
+		rb.Set = repoBranches
 
-	return repoBranches, r.err
+		return r.err
+	case 2:
+		// Length
+		l := r.uvarint()
+		branchIDs := make(map[string]*roaring.Bitmap, l)
+
+		for i := 0; i < l; i++ {
+			branch := r.str()
+			branchIDs[branch] = r.bitmap()
+		}
+
+		rb.IDs = branchIDs
+
+		return r.err
+	default:
+		return fmt.Errorf("unsupported RepoBranches encoding version %d", v)
+	}
 }
 
 type binaryReader struct {
 	b   []byte
+	off int
 	err error
 }
 
 func (b *binaryReader) uvarint() int {
-	x, n := binary.Uvarint(b.b)
+	x, n := binary.Uvarint(b.b[b.off:])
 	if n < 0 {
 		b.b = nil
 		b.err = errors.New("malformed RepoBranches")
 		return 0
 	}
-	b.b = b.b[n:]
+	b.off += n
 	return int(x)
 }
 
 func (b *binaryReader) str() string {
 	l := b.uvarint()
-	if l > len(b.b) {
+	if l > len(b.b[b.off:]) {
 		b.b = nil
 		b.err = errors.New("malformed RepoBranches")
 		return ""
 	}
-	s := b2s(b.b[:l])
-	b.b = b.b[l:]
+	s := b2s(b.b[b.off : b.off+l])
+	b.off += l
 	return s
 }
 
+func (b *binaryReader) bitmap() *roaring.Bitmap {
+	l := b.uvarint()
+	if l > len(b.b[b.off:]) {
+		b.b = nil
+		b.err = errors.New("malformed RepoBranches")
+		return nil
+	}
+	r := roaring.New()
+	_, b.err = r.FromBuffer(b.b[b.off : b.off+l])
+	b.off += l
+	return r
+}
+
 func (b *binaryReader) byt() byte {
-	if len(b.b) < 1 {
+	if len(b.b[b.off:]) < 1 {
 		b.b = nil
 		b.err = errors.New("malformed RepoBranches")
 		return 0
 	}
-	x := b.b[0]
-	b.b = b.b[1:]
+	x := b.b[b.off]
+	b.off++
 	return x
 }