This PR introduces a space optimized alternative to `query.RepoBranches`. It's name is `query.BranchesRepos` and it is a slice of `(Branch, Repo IDs bitmap)`. This is part of the effort to reduce memory spikes and OOMs in the sourcegraph frontend.