@tsenart 2021-09-27
Marshalling benchmarks looking good:

```
name                        time/op
RepoBranches_Encode/Set-12   655µs ±43%
RepoBranches_Encode/IDs-12  2.02µs ± 2%
RepoBranches_Decode/Set-12   917µs ± 1%
RepoBranches_Decode/IDs-12  15.4µs ± 1%

name                        bytes
RepoBranches_Encode/Set-12   344kB ± 0%
RepoBranches_Encode/IDs-12  1.75kB ± 0%
RepoBranches_Decode/Set-12
RepoBranches_Decode/IDs-12

name                        alloc/op
RepoBranches_Encode/Set-12   344kB ± 0%
RepoBranches_Encode/IDs-12  2.78kB ± 0%
RepoBranches_Decode/Set-12  1.44MB ± 0%
RepoBranches_Decode/IDs-12  11.1kB ± 0%

name                        allocs/op
RepoBranches_Encode/Set-12    2.00 ± 0%
RepoBranches_Encode/IDs-12    6.00 ± 0%
RepoBranches_Decode/Set-12     380 ± 0%
RepoBranches_Decode/IDs-12     194 ± 0%
```