@rmmh 2021-09-27
The core idea of this PR is changing `query.RepoBranches{Set: map[string][]string}` into `query.RepoBranches{Set: map[string]roaring.Bitmap}` and `query.RepoSet{Set: map[string]bool}` into `query.RepoSet{IDs: *roaring.Bitmap}`, right? This has a chance of returning results from the wrong branch or repo on a hash collision, will the frontend perform the correct filtering in that case?