@tsenart 2021-09-28
`RepoBranches` benchmarks with 5.5M repos

```
name                       time/op
RepoBranches_Encode-12      1.34s ±13%
RepoBranches_Decode-12      1.21s ± 2%
RepoBranchesIDs_Encode-12  21.7ms ±40%
RepoBranchesIDs_Decode-12  16.6ms ± 2%

name                       bytes
RepoBranches_Encode-12      189MB ± 0%
RepoBranches_Decode-12
RepoBranchesIDs_Encode-12  12.8MB ± 0%
RepoBranchesIDs_Decode-12

name                       alloc/op
RepoBranches_Encode-12      189MB ± 0%
RepoBranches_Decode-12      766MB ± 0%
RepoBranchesIDs_Encode-12  14.2MB ± 0%
RepoBranchesIDs_Decode-12  33.1MB ± 0%

name                       allocs/op
RepoBranches_Encode-12       1.00 ± 0%
RepoBranches_Decode-12       134k ± 0%
RepoBranchesIDs_Encode-12    5.00 ± 0%
RepoBranchesIDs_Decode-12    172k ± 0%
```