@tsenart 2021-09-28
> The core idea of this PR is changing query.RepoBranches{Set: map[string][]string} into query.RepoBranches{Set: map[string]roaring.Bitmap} and query.RepoSet{Set: map[string]bool} into query.RepoSet{IDs: *roaring.Bitmap}, right?

Yep!

> This has a chance of returning results from the wrong branch or repo on a hash collision, will the frontend perform the correct filtering in that case?

Not sure I understand. Can you elaborate on what you mean by a hash collision?