@tsenart 2021-09-28
Alright, last commit, now with RepoSet benchmarks too:

```
name                       time/op
RepoBranches_Encode-12      1.37s ± 7%
RepoBranches_Decode-12      1.38s ±32%
RepoBranchesIDs_Encode-12  26.1ms ± 4%
RepoBranchesIDs_Decode-12  21.2ms ±23%
RepoSet_Encode-12           1.36s ±11%
RepoSet_Decode-12           1.55s ± 1%
RepoSetIDs_Encode-12       17.9ms ± 3%
RepoSetIDs_Decode-12       14.9ms ± 3%

name                       bytes
RepoBranches_Encode-12      189MB ± 0%
RepoBranches_Decode-12
RepoBranchesIDs_Encode-12  12.8MB ± 0%
RepoBranchesIDs_Decode-12
RepoSet_Encode-12           187MB ± 0%
RepoSet_Decode-12
RepoSetIDs_Encode-12       11.5MB ± 0%
RepoSetIDs_Decode-12

name                       alloc/op
RepoBranches_Encode-12      189MB ± 0%
RepoBranches_Decode-12      766MB ± 0%
RepoBranchesIDs_Encode-12  14.2MB ± 0%
RepoBranchesIDs_Decode-12  33.1MB ± 0%
RepoSet_Encode-12           226MB ± 0%
RepoSet_Decode-12           536MB ± 0%
RepoSetIDs_Encode-12       33.6MB ± 0%
RepoSetIDs_Decode-12       26.3MB ± 0%

name                       allocs/op
RepoBranches_Encode-12       1.00 ± 0%
RepoBranches_Decode-12       134k ± 0%
RepoBranchesIDs_Encode-12    5.00 ± 0%
RepoBranchesIDs_Decode-12    172k ± 0%
RepoSet_Encode-12           11.0M ± 0%
RepoSet_Decode-12           5.52M ± 0%
RepoSetIDs_Encode-12         8.00 ± 0%
RepoSetIDs_Decode-12         131k ± 0%
```