@tsenart 2021-09-28
The PR should be a lot more "additive", so it's easier to review and later easier to delete deprecated code (i.e. `(RepoBranches
|RepoSet).Set)`)