@tsenart 2021-09-28
@keegancsmith: Yep, sequential IDs = massive gains, but more allocs. At some point we might want to contribute some optimizations on that front to roaring.

```
name                       old time/op    new time/op    delta
RepoBranches_Encode-12        1.37s ± 7%     1.37s ±13%       ~     (p=0.841 n=5+5)
RepoBranches_Decode-12        1.38s ±32%     1.27s ±17%       ~     (p=0.310 n=5+5)
RepoBranchesIDs_Encode-12    26.1ms ± 4%     0.2ms ± 9%    -99.35%  (p=0.008 n=5+5)
RepoBranchesIDs_Decode-12    21.2ms ±23%     0.3ms ±11%    -98.58%  (p=0.008 n=5+5)
RepoSet_Encode-12             1.36s ±11%     1.27s ±20%       ~     (p=0.151 n=5+5)
RepoSet_Decode-12             1.55s ± 1%     1.47s ±19%       ~     (p=0.190 n=4+5)
RepoSetIDs_Encode-12         17.9ms ± 3%     0.0ms ± 5%    -99.92%  (p=0.008 n=5+5)
RepoSetIDs_Decode-12         14.9ms ± 3%     0.0ms ± 2%    -99.73%  (p=0.008 n=5+5)

name                       old bytes      new bytes      delta
RepoBranches_Encode-12        189MB ± 0%     189MB ± 0%       ~     (all equal)
RepoBranchesIDs_Encode-12    12.8MB ± 0%     0.4MB ± 0%    -96.54%  (p=0.008 n=5+5)
RepoSet_Encode-12             187MB ± 0%     187MB ± 0%       ~     (all equal)
RepoSetIDs_Encode-12         11.5MB ± 0%     0.0MB ± 0%    -99.85%  (p=0.008 n=5+5)

name                       old alloc/op   new alloc/op   delta
RepoBranches_Encode-12        189MB ± 0%     189MB ± 0%       ~     (all equal)
RepoBranches_Decode-12        766MB ± 0%     766MB ± 0%       ~     (p=0.841 n=5+5)
RepoBranchesIDs_Encode-12    14.2MB ± 0%     0.5MB ± 0%       ~     (p=0.079 n=4+5)
RepoBranchesIDs_Decode-12    33.1MB ± 0%     0.9MB ± 0%    -97.22%  (p=0.008 n=5+5)
RepoSet_Encode-12             226MB ± 0%     226MB ± 0%       ~     (p=0.444 n=5+5)
RepoSet_Decode-12             536MB ± 0%     536MB ± 0%       ~     (p=0.421 n=5+5)
RepoSetIDs_Encode-12         33.6MB ± 0%     0.0MB ± 0%       ~     (p=0.079 n=4+5)
RepoSetIDs_Decode-12         26.3MB ± 0%     0.0MB ± 0%    -99.82%  (p=0.008 n=5+5)

name                       old allocs/op  new allocs/op  delta
RepoBranches_Encode-12         1.00 ± 0%      1.00 ± 0%       ~     (all equal)
RepoBranches_Decode-12         134k ± 0%      134k ± 0%       ~     (p=0.841 n=5+5)
RepoBranchesIDs_Encode-12      5.00 ± 0%     89.00 ± 0%  +1680.00%  (p=0.008 n=5+5)
RepoBranchesIDs_Decode-12      172k ± 0%        0k ± 0%    -99.74%  (p=0.008 n=5+5)
RepoSet_Encode-12             11.0M ± 0%     11.0M ± 0%       ~     (all equal)
RepoSet_Decode-12             5.52M ± 0%     5.52M ± 0%       ~     (p=0.421 n=5+5)
RepoSetIDs_Encode-12           8.00 ± 0%     87.00 ± 0%   +987.50%  (p=0.008 n=5+5)
RepoSetIDs_Decode-12           131k ± 0%        0k ± 0%    -99.66%  (p=0.008 n=5+5)
```