@rmmh 2021-09-28
As we discussed, the FNV hashing is only used for testing, otherwise we use repo IDs from our SQL backend.

The IDs generated for benchmarking shouldn't be *entirely* sequential, right? There are some gaps in the repo ids table (latest repoid is 56M but there are 2M active), so maybe make the generator have a random spacing in [0, 50]? The additional allocs are trivial and unconcerning-- an extra 80B of allocs here doesn't matter at all.