@stefanhengl 2021-09-29
I was actually hoping we move away from RepoSets and RepoBranches, i.e. sending lists of repos around, and instead let Zoekt handle the raw query (for lack of a better word). Is there any reason why we absolutely have to send lists of repositories to Zoekt? @rvantonder is working on translating queries to Zoekt queries early, maybe we can build on that?