@keegancsmith 2021-09-29
Did some faking so we can run benchstat to compare RepoBranches vs the new BranchesRepos:
```
name                     old time/op    new time/op    delta
BranchesRepos_Encode-32     1.03s ± 7%     0.00s ± 3%    -99.99%  (p=0.008 n=5+5)
BranchesRepos_Decode-32     915ms ± 4%       0ms ± 1%    -99.98%  (p=0.008 n=5+5)

name                     old bytes      new bytes      delta
BranchesRepos_Encode-32     189MB ± 0%       0MB ± 0%    -99.77%  (p=0.008 n=5+5)

name                     old alloc/op   new alloc/op   delta
BranchesRepos_Encode-32     189MB ± 0%       0MB ± 0%    -99.76%  (p=0.000 n=4+5)
BranchesRepos_Decode-32     766MB ± 0%       1MB ± 0%    -99.88%  (p=0.008 n=5+5)

name                     old allocs/op  new allocs/op  delta
BranchesRepos_Encode-32      1.60 ±88%     89.00 ± 0%  +5462.50%  (p=0.008 n=5+5)
BranchesRepos_Decode-32      134k ± 0%        0k ± 0%    -99.67%  (p=0.008 n=5+5)
```

You can now pick up your jaw off the floor.